Enable team creation at user creation
User can create a new team when signing up, or associate with an already
existing team. Will need to be styled, and ideally made so that only one
of the two forms is available at a time.

As is it is possible to have
both forms displayed and filled out. For the time being we assume that
if both are present the new team field is used. Making it so that only
one is available at a time would remove this amniguation.