Add number of questions answered to team model
Can use in the future to display the current leader. Need to figure out
a clean way to get the current leader, and in the case of ties, display
that there is no leader. Also need to make sure it doesn't break when
there is only one team, during setup.