Add tracking of correct answers
Answers are now marked as answered for a given team when a user answers
the question correctly. Warnings are provided when they are looking at a
question their team has already answered, and they are told why they
cannot answer a question (the submit button is removed) when they are
not logged in, and are browsing through questions.

Really coming together!