Display errors caused by invalid user creation
Currently assigning these errors to an array in the flash, but that
makes it harder to display. I'd rather not have to list that specific
hash key in the flash as an exception in the site wide header, but if
that's what I have to do I guess it's not the end of the world.